Revert "test"
This reverts commit 88500f4eaec0d14d432a531fa8730c8baf5f290d.

+++ BLL/Controller/PoddController.cs
@@ -66,14 +66,38 @@ namespace BLL.Controller
 
         public List<Pod> GetAllPods()
         {
-            return podRepository.GetAll();
+            return podRepository.GetAll().Distinct().ToList();
         }
 
 
         public void DeletePod(string name)
         {
             int index = podRepository.GetIndex(name);
-            podRepository.Delete(index);
+            if(index >= 0)
+            {
+                var podToRemove = podRepository.GetByName(name);
+                podRepository.Delete(index);
+
+                var kategori = kategoriController.LasAllaKategorier()
+                .Select(kName => kategoriController.hamtaKategoriByName(kName))
+                .FirstOrDefault(k => k != null && k.Pod != null && k.Pod.Any(p => p.PodTitel.Equals(name, StringComparison.OrdinalIgnoreCase)));
+
+                if (kategori != null)
+                {
+                    var poddToRemove = kategori.Pod.FirstOrDefault(p => p.PodTitel.Equals(name, StringComparison.OrdinalIgnoreCase));
+                    if (poddToRemove != null)
+                    {
+                        kategori.Pod.Remove(poddToRemove);
+                        kategoriController.UpdateraKategori(kategori);
+                    }
+                }
+
+            }
+            else
+            {
+                Console.WriteLine("Podden kunde inte hittas och togs därför inte bort.");
+            }
+            
         }
 
         //public string GetPodName(string name)
@@ -101,8 +125,13 @@ namespace BLL.Controller
             }
 
             Pod nyPod = new Pod {PodUrl = rssLank, PodTitel = name};
-            kategori.Pod.Add(nyPod);
-            kategoriController.UpdateraKategori(kategori);
+
+            if(!podRepository.GetAll().Any(p => p.PodTitel.Equals(nyPod.PodTitel, StringComparison.OrdinalIgnoreCase)))
+            {
+                kategori.Pod.Add(nyPod);
+                podRepository.Create(nyPod);
+                kategoriController.UpdateraKategori(kategori);
+            }
         }
         public Pod GetPodByTitle(string title)
         {
